+++ src/Eve-O-Preview/Configuration/Interface/IThumbnailConfiguration.cs
@@ -22,6 +22,7 @@ namespace EveOPreview.Configuration
 		bool EnableClientLayoutTracking { get; set; }
 		bool HideActiveClientThumbnail { get; set; }
 		bool ShowThumbnailPreviews { get; set; }
+		bool PreviewMinimizedClients { get; set; }
 		bool HideLoginClientThumbnail { get; set; }
 		bool MinimizeInactiveClients { get; set; }
 		bool HideCaptionOnClients { get; set; }
@@ -62,6 +63,7 @@ namespace EveOPreview.Configuration
 
 		string IconName { get; set; }
 		List<string> MinimizeAllClientsHotkeys { get; set; }
+		List<string> ShowAllPreviewsHotkeys { get; set; }
 
 		Point LoginThumbnailLocation { get; set; }
 
@@ -77,6 +79,7 @@ namespace EveOPreview.Configuration
 		void SetClientHotkey(string currentClient, Keys hotkey);
 		Keys StringToKey(string hotkey);
 		bool IsPriorityClient(string currentClient);
+		void SetPriorityClient(string currentClient, bool isPriority);
 		IReadOnlyList<string> ExecutablesToPreview { get; }
 
 		bool IsExecutableToPreview(string processName);