+++ src/Eve-O-Preview/Services/Interface/InteropConstants.cs
@@ -78,6 +78,7 @@ namespace EveOPreview.Services
 		public const int SW_SHOWNORMAL = 1;
 		public const int SW_SHOWMINIMIZED = 2;
 		public const int SW_SHOWMAXIMIZED = 3;
+		public const int SW_SHOWNOACTIVATE = 4;
 		public const int SW_RESTORE = 9;
 	}
 }